Let clang-tidy look for defects rather than opinions
Two checks were on out of about five hundred. Turning the rest on and
counting says why it stayed that way: 16248 warnings across 86 files,
and almost all of it argues with the house style rather than finding
anything. 4134 want trailing return types, 2583 want nodiscard, 2510 are
the *Data structs whose members are public because glaze reflects over
them, 2043 want 1.0F instead of 1.0f, 2432 across two checks call a
game's tuning constants magic numbers, 624 want braces the codebase
deliberately leaves off, and 367 are glm's own unions.

The defects live in the two smallest groups. clang-analyzer found five
warnings and every one is real. bugprone found sixty-five. Those two,
plus performance, are now on.

What the five were

Four of them are the tile grid: a float loop counter stepping x += spacing
across the screen and accumulating as it goes, in the file that had
already been reported as looking wrong. The fifth is the same pattern
scanning for jump take-offs. Both loops count in ints now and multiply,
so the thousandth line is exactly a thousand spacings from the first, and
a test walks five thousand of them to say so. The grid also stopped
treating tile coordinates as floats: tileX and tileY were floats used to
index a tile map and formatted into a label, narrowed on every use.

What bugprone found worth keeping

Four unchecked optional accesses, all the same shape: a caller checks and
the callee dereferences again, so the precondition lives in a comment
nobody wrote. FleeBehavior::furthestAlong and RouteWalker::targetPosition
take the unwrapped values now, and the check is where the answer is.

Two string_view::data() handed to ImGui where a null terminated string is
wanted. Making the section names const char * would have settled it and
crashed the macOS compiler instead: a frontend segfault on the two files
that read them, on the same clang that builds them here without
complaint. They stay string_view and say in a NOLINT that the views are
over string literals. A move of a const expression
that copies anyway. Four std::endl on paths that are about to return.
A conditional whose two arms did the same thing. Two vectors filled from
a container of known size without reserving.

What was turned back off

bugprone-easily-swappable-parameters, eighteen of them and no defect
among them. performance-enum-size, eighty-seven asking for smaller
underlying types. And the integer to floating point half of
narrowing-conversions: twenty-one of its twenty-two were int to float,
which is exact for every value here, and the one that was not is
unsigned long to int and is fixed.

An assert is not a check in the build that ships. tile_interaction_system
asserted a pickup optional had a value and then read it, and CI lints a
Release configure where NDEBUG deletes the assert and leaves the read.
Found by the ci that this change turned on, missed here because Debug was
what got linted. It is an if now.

Tests get their own config. What fires there is the framework: TEST_CASE
expands to a static object 399 times, and a REQUIRE before dereferencing
an optional is a guarantee no analysis can see. They keep the naming and
include checks, which is what a test can still get wrong on its own.

+++ src/serialization/json_format.cpp
@@ -115,7 +115,7 @@ std::string withStructureOnLines(const std::string &json)
             expanded.push_back(onLines);
 
             if (onLines)
-                out += "\n" + std::string(4 * depth, ' ');
+                out += "\n" + std::string(static_cast<std::size_t>(4 * depth), ' ');
         }
         else if (character == '}' || character == ']')
         {
@@ -125,7 +125,7 @@ std::string withStructureOnLines(const std::string &json)
 
             --depth;
             if (onLines)
-                out += "\n" + std::string(4 * depth, ' ');
+                out += "\n" + std::string(static_cast<std::size_t>(4 * depth), ' ');
 
             out += character;
         }
@@ -133,7 +133,7 @@ std::string withStructureOnLines(const std::string &json)
         {
             out += character;
             if (!expanded.empty() && expanded.back())
-                out += "\n" + std::string(4 * depth, ' ');
+                out += "\n" + std::string(static_cast<std::size_t>(4 * depth), ' ');
         }
         else
         {
